+++ basic/25-multi-sig-wallet/hardhat.config.js
@@ -46,23 +46,23 @@ module.exports = {
         (you can put in a mnemonic here to set the deployer locally)
       */
     },
-    goerli: {
-      url: "https://goerli.infura.io/v3/" + process.env.INFURA_ID, //<---- YOUR INFURA ID! (or it won't work)
-      accounts: [
-        mnemonic()
-      ],
-    },
-    mainnet: {
-      url: "https://mainnet.infura.io/v3/" + process.env.INFURA_ID, //<---- YOUR INFURA ID! (or it won't work)
-      accounts: [
-        mnemonic()
-      ],
-    },
-    ropsten: {
-      url: "https://ropsten.infura.io/v3/" + process.env.INFURA_ID, //<---- YOUR INFURA ID! (or it won't work)
-      accounts: [
-        mnemonic()
-      ],
-    },
+    // goerli: {
+    //   url: "https://goerli.infura.io/v3/" + process.env.INFURA_ID, //<---- YOUR INFURA ID! (or it won't work)
+    //   accounts: [
+    //     mnemonic()
+    //   ],
+    // },
+    // mainnet: {
+    //   url: "https://mainnet.infura.io/v3/" + process.env.INFURA_ID, //<---- YOUR INFURA ID! (or it won't work)
+    //   accounts: [
+    //     mnemonic()
+    //   ],
+    // },
+    // ropsten: {
+    //   url: "https://ropsten.infura.io/v3/" + process.env.INFURA_ID, //<---- YOUR INFURA ID! (or it won't work)
+    //   accounts: [
+    //     mnemonic()
+    //   ],
+    // },
   }
 };
